LayoutTests/imported/w3c/web-platform-tests/IndexedDB/bigint_value-expected.txt

CONSOLE MESSAGE: line 41: SyntaxError: No identifiers allowed directly after numeric literal

FAIL IndexedDB: BigInt keys and values SyntaxError: No identifiers allowed directly after numeric literal
